pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "folium>=0.20.0",
    "googletrans>=4.0.2",
    "numpy==1.26.4",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "schedule>=1.2.2",
    "streamlit-folium>=0.25.0",
    "streamlit>=1.46.0",
    "trafilatura>=2.0.0",
    "twilio>=9.6.3",
    "plotly>=6.1.2",
    "scikit-learn>=1.7.0",
    "feedparser>=6.0.11",
]
